==> Stars - Health Plans/09_summary_stars.Rmd <==
---
title: "Summary/Overall Star Scores"
output:
html_notebook: default
---


```{r}
source("init.R")

load(file.path(dirs$rdata, "08_contract_df.RData"))
```


```{r}
file_vec <- get_unzipped_files_by_pattern("summary.*csv$")
file_df <-
  tibble(
    star_year = as.integer(str_sub(names(file_vec), end = 4L)),
    file_name = unname(file_vec)
  ) %>%
  mutate(
    part_c_only = grepl("part.?c[^/]*$", file_name, ignore.case = TRUE),
    part_d_only = grepl("part.?d[^/]*$", file_name, ignore.case = TRUE),
    fall   = grepl("fall.?release", file_name, ignore.case = TRUE),
    spring = grepl("spring.?release", file_name, ignore.case = TRUE)
  ) %>%
  relocate(file_name, .after = spring) %>%
  filter(!spring) %>%
  mutate(file_key = str_c(star_year,
                          ifelse(part_c_only, "C", ""),
                          ifelse(part_d_only, "D", ""))) %>%
  relocate(file_key, .before = star_year) %>%
  inner_join(
    tribble(
      ~file_key, ~skip_rows, ~col_ind,   ~snp_col,    ~disaster_cols,
    	"2009C",   1L,         c(1, 5),    NA_integer_, c(),
    	"2009D",   2L,         c(1, 5),    NA_integer_, c(),
    	"2010C",   1L,         c(1, 5),    NA_integer_, c(),
    	"2010D",   3L,         c(1, 4),    NA_integer_, c(),
    	"2011C",   1L,         c(1, 6, 7), NA_integer_, c(),
    	"2011D",   3L,         c(1, 4),    NA_integer_, c(),
    	"2012C",   1L,         c(1, 5, 8), NA_integer_, c(),
    	"2012D",   3L,         c(1, 6),    NA_integer_, c(),
    	"2013C",   1L,         c(1, 6, 9), NA_integer_, c(),
    	"2013D",   3L,         c(1, 6),    NA_integer_, c(),
    	"2014C",   1L,         c(1, 8, 9), NA_integer_, c(),
    	"2014D",   1L,         c(1, 8),    NA_integer_, c(),
    	"2015",    1L,         c(1, 8:10), NA_integer_, c(),
    	"2016",    1L,         c(1, 8:10), NA_integer_, c(),
    	"2017",    1L,         c(1, 8:10), NA_integer_, c(),
    	"2018",    1L,         c(1, 7:9),  NA_integer_, c(),
    	"2019",    1L,         c(1, 7:9),  6L,          c(),
    	"2020",    1L,         c(1, 9:11), 6L,          c(7,8),
    	"2021",    1L,         c(1, 9:11), 6L,          c(7,8),
    	"2022",    1L,         c(1, 9:11), 6L,          c(7,8),
      "2023",    1L,         c(1, 9:11), 6L,          c(7,8),
      "2024",    1L,         c(1, 9:11), 6L,          c(7,8),
    ),
    by = "file_key"
  )


process_one_summary_file <- function(file_name, skip_rows, col_ind, part_c_only) {
  df <- read.csv(file_name, skip = skip_rows, check.names = FALSE)[, col_ind] %>%
    mutate(across(everything(), str_trim))
  all_cnames <- c("contract_id", "Part C Summary",
                  "Part D Summary", "Overall Summary")
  part_col_ind <- ifelse(part_c_only, 2, 3)
  if (ncol(df) == 2) {
    colnames(df) <- all_cnames[c(1, part_col_ind)]
  } else if(ncol(df) == 3) {
    colnames(df) <- all_cnames[c(1, part_col_ind, 4)]
  } else if(ncol(df) == 4) {
    colnames(df) <- all_cnames
  }
  df %>%
    pivot_longer(all_of(colnames(df)[-1]),
                 names_to = "summary_type", values_to = "stars")
}

summary_stars_df <-
  file_df %>%
  mutate(
    star_year = as.integer(star_year),
    summary_df = pmap(list(file_name, skip_rows, col_ind, part_c_only),
                      process_one_summary_file)
  ) %>%
  select(star_year, summary_df) %>%
  unnest(summary_df) %>%
  filter(str_trim(stars) != "", str_trim(contract_id) != "") %>%
  mutate(
    stars = str_trim(gsub("( out of 5)? stars$", "", stars)),
    stars_note = ifelse(!grepl("^\\d(.\\d)?$", stars), stars, NA_character_),
    stars = ifelse(grepl("^\\d(.\\d)?$", stars), as.numeric(stars), NA_real_)
  ) %>%
  transmute(
    star_year,
    summary_type = case_when(summary_type == "Part C Summary"  ~ "Part C",
                             summary_type == "Part D Summary"  ~ "Part D",
                             summary_type == "Overall Summary" ~ "Overall"),
    contract_id,
    stars,
    stars_note
  )

summary_stars_df
```



```{r}
process_one_summary_file_snp_disaster <-
  function(file_name, skip_rows, snp_col, disaster_cols) {
    
  col_ind <- c(1, snp_col, disaster_cols)
  df <-
    read.csv(file_name, skip = skip_rows, check.names = FALSE)[, col_ind] %>%
    mutate(across(everything(), str_trim))
  all_cnames <- c("contract_id", "snp_flag")
  if (length(disaster_cols) == 2) {
    all_cnames <-
      c(all_cnames, "disaster_pcg_year_minus_3", "disaster_pcg_year_minus_2")
  }
  colnames(df) <- all_cnames
  df
}


snp_disaster_df <-
  file_df %>%
  mutate(disaster_len = map_int(disaster_cols, length)) %>%
  filter(!is.na(snp_col) | disaster_len > 0) %>%
  mutate(df = pmap(list(file_name, skip_rows, snp_col, disaster_cols),
                   process_one_summary_file_snp_disaster)) %>%
  select(star_year, df) %>%
  unnest(df) %>%
  mutate(
    snp_flag = (snp_flag == "Yes"),
    disaster_pcg_year_minus_3 = as.numeric(disaster_pcg_year_minus_3) / 100,
    disaster_pcg_year_minus_2 = as.numeric(disaster_pcg_year_minus_2) / 100
  )

contract_df <-
  contract_df %>%
  left_join(snp_disaster_df, by = c("star_year", "contract_id"))

contract_df
```





```{r}
save(summary_stars_df,
     file = file.path(dirs$rdata, "09_summary_stars_df.RData"))
save(contract_df, file = file.path(dirs$rdata, "09_contract_df.RData"))
```








# Testing and scratch work



```{r, eval=FALSE}
file_df
file_df %>% filter(!part_c_only, !part_d_only) %>% pull(file_name)
file_df %>% filter(!fall, !spring)
file_df %>% filter(fall, spring)
(2008:2022)[!(as.character(2008:2022) %in% file_df$star_year)]
file_df %>% count(star_year, part_c_only, part_d_only, fall, spring) %>% filter(n > 1)
file_df %>% filter(!spring) %>% count(star_year, part_c_only, part_d_only) %>% filter(n > 1)

file_df
```


```{r, eval=FALSE}
read.csv(file_df[file_df$file_key == "2009C", ]$file_name, skip = 1, check.names = FALSE)[, c(1, 5)]
read.csv(file_df[file_df$file_key == "2009D", ]$file_name, skip = 2, check.names = FALSE)[, c(1, 5)]
read.csv(file_df[file_df$file_key == "2010C", ]$file_name, skip = 1, check.names = FALSE)[, c(1, 5)]
read.csv(file_df[file_df$file_key == "2010D", ]$file_name, skip = 3, check.names = FALSE)[, c(1, 4)]
read.csv(file_df[file_df$file_key == "2011C", ]$file_name, skip = 1, check.names = FALSE)[, c(1, 6, 7)]
read.csv(file_df[file_df$file_key == "2011D", ]$file_name, skip = 3, check.names = FALSE)[, c(1, 4)]
read.csv(file_df[file_df$file_key == "2012C", ]$file_name, skip = 1, check.names = FALSE)[, c(1, 5, 8)]
read.csv(file_df[file_df$file_key == "2012D", ]$file_name, skip = 3, check.names = FALSE)[, c(1, 6)]
read.csv(file_df[file_df$file_key == "2013C", ]$file_name, skip = 1, check.names = FALSE)[, c(1, 6, 9)]
read.csv(file_df[file_df$file_key == "2013D", ]$file_name, skip = 3, check.names = FALSE)[, c(1, 6)]
read.csv(file_df[file_df$file_key == "2014C", ]$file_name, skip = 1, check.names = FALSE)[, c(1, 8, 9)]
read.csv(file_df[file_df$file_key == "2014D", ]$file_name, skip = 1, check.names = FALSE)[, c(1, 8)]
read.csv(file_df[file_df$file_key == "2015", ]$file_name, skip = 1, check.names = FALSE)[, c(1, 8:10)]
read.csv(file_df[file_df$file_key == "2016", ]$file_name, skip = 1, check.names = FALSE)[, c(1, 8:10)]
read.csv(file_df[file_df$file_key == "2017", ]$file_name, skip = 1, check.names = FALSE)[, c(1, 8:10)]
read.csv(file_df[file_df$file_key == "2018", ]$file_name, skip = 1, check.names = FALSE)[, c(1, 7:9)]
read.csv(file_df[file_df$file_key == "2019", ]$file_name, skip = 1, check.names = FALSE)[, c(1, 7:9)]
read.csv(file_df[file_df$file_key == "2020", ]$file_name, skip = 1, check.names = FALSE)[, c(1, 9:11)]
read.csv(file_df[file_df$file_key == "2021", ]$file_name, skip = 1, check.names = FALSE)[, c(1, 9:11)]
read.csv(file_df[file_df$file_key == "2022", ]$file_name, skip = 1, check.names = FALSE)[, c(1, 9:11)]

# Columns I ignored:
#   * SNP
#   * 2017 Disaster %, 2018 Disaster %, 2019 Disaster %, 2020 Disaster %
```


```{r, eval=FALSE}


temp_df %>% count(stars) %>% arrange(desc(n))

temp_df %>% filter(stars == "")
temp_df %>% filter(contract_id == "")

summary_df
```
