==> Stars - Health Plans/10_cai_files.Rmd <==
---
title: "R Notebook"
output:
html_notebook: default
---



```{r}
source("init.R")

load(file.path(dirs$rdata, "09_contract_df.RData"))
cai_val_df <- read.csv(file.path(dirs$table, "cai_val_df.csv")) %>% as_tibble()
```


*TODO*: Add `Puerto Rico Only` flag from CAI files to `contract_df`

```{r}
cai_file_vec <- get_unzipped_files_by_pattern("cai.*csv$")

cai_file_df <-
  tibble(
    star_year = str_sub(names(cai_file_vec), end = 4L),
    file_name = unname(cai_file_vec)
  ) %>%
  mutate(
    part_c_only = grepl("part.?c[^/]*$", file_name, ignore.case = TRUE),
    part_d_only = grepl("part.?d[^/]*$", file_name, ignore.case = TRUE),
    fall   = grepl("fall.?release", file_name, ignore.case = TRUE),
    spring = grepl("spring.?release", file_name, ignore.case = TRUE)
  ) %>%
  relocate(file_name, .after = spring) %>%
  filter(!spring) %>%
  mutate(file_key = str_c(star_year,
                          ifelse(part_c_only, "C", ""),
                          ifelse(part_d_only, "D", ""))) %>%
  relocate(file_key, .before = star_year)


process_one_cai_file <- function(file_name, skip_rows, col_ind) {
  df <-
    read.csv(file_name, skip = skip_rows, check.names = FALSE)[, col_ind] %>%
    mutate(across(everything(), str_trim))
  colnames(df) <- c("contract_id", "Part C",
                    "Part D MA-PD", "Part D PDP", "Overall")
  df %>%
    pivot_longer(all_of(colnames(df)[-1]),
                 names_to = "summary_type", values_to = "fac")
}


cai_df <-
  tribble(
    ~file_key, ~skip_rows, ~col_ind,
  	"2017",    1,          c(1, 6:9),
  	"2018",    1,          c(1, 6:9),
  	"2019",    1,          c(1, 6:9),
  	"2020",    1,          c(1, 6:9),
  	"2021",    1,          c(1, 6:9),
  	"2022",    1,          c(1, 6:9),
  	"2023",    1,          c(1, 6:9),
  	"2024",    1,          c(1, 6:9),
  ) %>%
  inner_join(cai_file_df, by = "file_key") %>%
  mutate(
    star_year = as.integer(star_year),
    cai_df = pmap(list(file_name, skip_rows, col_ind),
                      process_one_cai_file)
  ) %>%
  select(star_year, cai_df) %>%
  unnest(cai_df) %>%
  filter(!is.na(fac), fac != "N/A") %>%
  select(star_year, contract_id, summary_type, fac) %>%
  left_join(cai_val_df, by = c("star_year", "summary_type", "fac"))

cai_df
```


```{r}
process_one_cai_file_puerto_rico <- function(filename) {
  df <- read.csv(filename, skip = 1, check.names = FALSE)[, c(1,5)]
  colnames(df) <- c("contract_id", "puerto_rico_only")
  df %>%
    mutate(across(everything(), str_trim)) %>%
    mutate(puerto_rico_only = (puerto_rico_only == "Yes"))
}

puerto_rico_only_df <-
  cai_file_df %>%
  transmute(
    star_year = as.integer(star_year),
    df = map(file_name, process_one_cai_file_puerto_rico)
  ) %>%
  unnest(df)

min_star_year <- as.integer(min(cai_file_df$star_year))

contract_df <-
  contract_df %>%
  left_join(puerto_rico_only_df, by = c("star_year", "contract_id")) %>%
  mutate(
    puerto_rico_only = map2_lgl(
      star_year, puerto_rico_only,
      ~ifelse(as.integer(.x) >= min_star_year, replace_na(.y, FALSE), .y)
    )
  )

contract_df
```




```{r}
save(cai_df, file = file.path(dirs$rdata, "10_cai_df.RData"))
save(contract_df, file = file.path(dirs$rdata, "10_contract_df.RData"))
```




# Testing


```{r, eval=FALSE}
cai_df

cai_df %>%
  count(star_year, fac) %>%
  pivot_wider(names_from = star_year, values_from = n) %>%
  mutate(across(all_of(as.character(2017:2022)), function(x) replace_na(x, 0L)))
```

```{r, eval=FALSE}
cai_df %>% filter(is.na(cai))
```

```{r, eval=FALSE}
cai_df %>% count(star_year, contract_id, summary_type) %>% filter(n>1)
```
